Cleaned up reduction dispatch code and separated iterators into separate files
